requirements.txt
anthropic
imapclient
playwright
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic>=2.0
pydantic-settings
APScheduler
python-dotenv
structlog
tenacity
pypdf
